flask_AI/requirements.txt
kafka-python
pandas
scikit-learn
imbalanced-learn
urllib3
requests

